Move System.Net.RequestCachePolicy/Level down to System.Net.Primitives
This allows us to remove the System.Xml dependency on System.Net.Requests.
