+++ src/System.Private.Xml/src/project.json
@@ -27,7 +27,7 @@
         "System.IO.FileSystem.Primitives": "4.4.0-beta-24619-02",
         "System.Linq": "4.4.0-beta-24619-02",
         "System.Net.Http": "4.4.0-beta-24619-02",
-        "System.Net.Requests": "4.4.0-beta-24619-02",
+        "System.Net.Primitives": "4.4.0-beta-24619-02",
         "System.Reflection": "4.4.0-beta-24619-02",
         "System.Reflection.Emit": "4.4.0-beta-24619-02",
         "System.Reflection.Emit.ILGeneration": "4.4.0-beta-24619-02",
@@ -75,7 +75,7 @@
         "System.IO.FileSystem.Primitives": "4.4.0-beta-24619-02",
         "System.Linq": "4.4.0-beta-24619-02",
         "System.Net.Http": "4.4.0-beta-24619-02",
-        "System.Net.Requests": "4.4.0-beta-24619-02",
+        "System.Net.Primitives": "4.4.0-beta-24619-02",
         "System.Reflection": "4.4.0-beta-24619-02",
         "System.Reflection.Emit": "4.4.0-beta-24619-02",
         "System.Reflection.Emit.ILGeneration": "4.4.0-beta-24619-02",
