+++ src/System.Xml.ReaderWriter/ref/project.json
@@ -4,7 +4,6 @@
     "System.Diagnostics.Debug": "4.4.0-beta-24619-02",
     "System.IO": "4.4.0-beta-24619-02",
     "System.Net.Primitives": "4.4.0-beta-24619-02",
-    "System.Net.Requests": "4.4.0-beta-24619-02",
     "System.Runtime": "4.4.0-beta-24619-02",
     "System.Text.Encoding": "4.4.0-beta-24619-02",
     "System.Threading.Tasks": "4.4.0-beta-24619-02"
